A more compact and useful representation for unphased data (#13)
For unphased data, each variant stores the number of copies that it
represents (1...ploidy). E.g., for diploid data you'll have potentially
two copies of each variant, one with num_copies=1 and one with
num_copies=2 (homozygotes).

Included is an example that shows how to calculate ROH from this
format.